requirements.txt
absl-py==1.4.0
adafruit-ampy==1.1.0
aiohttp==3.9.1
aiomysql==0.2.0
aiosignal==1.3.1
altgraph==0.17.3
annotated-types==0.6.0
anyio==3.7.1
AnyQt==0.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.3
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
baycomp==1.0.2
bcrypt==4.0.1
beautifulsoup4==4.12.2
bitstring==4.0.2
bleach==6.0.0
bleak==0.20.2
bleak-winrt==1.2.0
blinker==1.6.2
Bootstrap-Flask==2.2.0
Bottleneck==1.3.7
branca==0.6.0
Brotli==1.1.0
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
comm==0.1.3
commonmark==0.9.1
comtypes==1.2.0
contourpy==1.1.0
cryptography==40.0.2
cssselect2==0.7.0
customtkinter==5.2.0
cycler==0.11.0
Cython==0.29.35
darkdetect==0.8.0
databases==0.8.0
dataclasses-json==0.6.3
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dictdiffer==0.9.0
distro==1.8.0
dnspython==2.4.2
docutils==0.20.1
docx2pdf==0.1.8
dominate==2.8.0
EasyProcess==1.1
ecdsa==0.18.0
email-validator==2.0.0.post2
entrypoint2==1.1
enum-compat==0.0.3
esptool==4.5.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.104.1
fastjsonschema==2.18.0
filelock==3.12.2
fiona==1.9.5
Flask==2.3.1
Flask-Material==0.1.1
flatbuffers==23.3.3
folium==0.14.0
fonttools==4.40.0
fqdn==1.5.1
frozenlist==1.4.0
gast==0.4.0
gatt==0.2.7
geopandas==0.14.0
gitdb==4.0.10
GitPython==3.1.32
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-api-python-client==2.90.0
google-auth==2.25.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-generativeai==0.3.1
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
greenlet==3.0.1
growattRS232==1.0.0
grpcio==1.60.0
grpcio-status==1.60.0
gspread==5.9.0
gunicorn==21.2.0
h11==0.14.0
h5py==3.8.0
html5lib==1.1
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.0
httpx==0.25.2
huami_token==0.7.0
idna==3.6
importlib-metadata==6.7.0
influxdb==5.3.1
influxdb-client==1.36.1
ipykernel==6.23.1
ipython==8.18.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
jax==0.4.8
jedi==0.19.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.23.0
keras==2.12.0
keyboard==0.13.5
keyring==24.0.1
keyrings.alt==4.2.0
kiwisolver==1.4.4
labelImg==1.8.6
langchain==0.0.346
langchain-core==0.0.10
langsmith==0.0.69
lazypredict==0.2.12
libclang==16.0.0
lightgbm==4.1.0
lxml==4.9.3
Markdown==3.4.3
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==3.0.1
ml-dtypes==0.1.0
more-itertools==9.1.0
MouseInfo==0.1.3
mpmath==1.3.0
mpy-cross==1.19.1
msgpack==1.0.5
mss==9.0.1
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.2.0
nbclient==0.8.0
nbconvert==7.7.2
nbformat==5.9.1
nest-asyncio==1.5.6
networkx==3.1
notebook==7.0.0
notebook_shim==0.2.3
numpy==1.26.2
oauthlib==3.2.2
openai==1.3.7
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
openpyxl==3.1.2
openTSNE==1.0.0
opt-einsum==3.3.0
orange-canvas-core==0.1.31
orange-widget-base==4.21.0
Orange3==3.35.0
orjson==3.9.7
overrides==7.3.1
packaging==23.2
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
patsy==0.5.3
pefile==2023.2.7
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
plotly==5.17.0
plum-dispatch==1.7.4
ply==3.11
pmdarima==2.0.4
prometheus-client==0.17.1
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.1
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.5.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.14.5
pydyf==0.8.0
pygame==2.3.0
pygatt==4.0.5
PyGetWindow==0.0.9
Pygments==2.17.2
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.6
pymodbus==3.3.2
pyModbusTCP==0.2.0
pymongo==4.6.1
PyMsgBox==1.0.9
PyMySQL==1.1.0
pyodbc==4.0.39
pyparsing==3.0.9
pyperclip==1.8.2
pyphen==0.14.0
pyproj==3.6.1
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyqtgraph==0.13.3
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
pyserial==3.5
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
python-louvain==0.16
python-multipart==0.0.6
python-time==0.3.0
python-vlc==3.0.18122
pytweening==1.0.7
pytz==2023.3
pyusb==1.2.1
pywin32==306
pywin32-ctypes==0.2.1
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.0
qasync==0.24.0
qtconsole==5.4.3
QtPy==2.3.1
reactivex==4.0.4
reedsolo==1.6.0
referencing==0.30.0
reportlab==4.0.7
requests==2.31.0
requests-oauthlib==1.3.1
rethinkdb==2.4.9
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.9.2
rsa==4.9
scikit-learn==1.1.3
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.2
serverfiles==0.3.1
shapely==2.0.2
sip==6.7.9
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
Spire.Doc==12.1.0
SQLAlchemy==2.0.23
stack-data==0.6.3
starlette==0.27.0
statsmodels==0.14.0
sympy==1.12
tenacity==8.2.3
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
thinker==1.1.1
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tinycss2==1.2.1
torch==2.0.1
torchaudio==2.0.2+cu117
torchvision==0.15.2
tornado==6.3.2
tqdm==4.66.1
traitlets==5.14.0
types-requests==2.25.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
ultralytics==8.0.141
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.24.0.post1
visitor==0.1.3
watchfiles==0.20.0
wcwidth==0.2.12
weasyprint==60.2
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==2.3.1
widgetsnbextension==4.0.9
wrapt==1.14.1
WTForms==3.0.1
xgboost==2.0.0
xlrd==2.0.1
XlsxWriter==3.1.2
yarl==1.9.4
zipp==3.15.0
zopfli==0.2.3
